Split HIR generic analysis module (#1297)
Move generic inference, trait-bound validation, and monomorphization out of hir.rs into a compiler.hir submodule. Update the compiler source monolith reporter and ratchet plan so the new module is tracked with lowered top-file ceilings.

Refs #1254.
